Link leaders and followers (and introduce follower_wait)
Adding links between leaders and followers, we can't deny access
for seconds.

We don't want links between followers because that would become
a fully connected mesh, which I think we want to avoid on this level.

Introduced a follower_wait role/state when we postpone events, a similar
role/state to candidate but for the non leaders.
It might be used while merging states as well when the network is not
available.

Problematic stuff in this and previous commits was discovered when
adding more test cases.